Finished

+++ part3.sh
@@ -1,7 +1,11 @@
 #!/bin/bash
-rm -rf /etc/rc.local
+if [ -f "/etc/rc.d/rc.local" ]; then
+  sed -i '/\/root\/part2\.sh/d' /etc/rc.d/rc.local
+  chmod -x /etc/rc.d/rc.local
+else 
+  rm -rf /etc/rc.local
+fi
 rm -rf /root/part3.sh
-rm -rf /etc/rc.d/rc.local
 sudo dd if=/dev/zero of=/EMPTY bs=1M
 sudo rm -f /EMPTY
 shutdown -h +0
